feat(document): add endpoint to fetch documents by workflow ID
- Implement new route for retrieving documents based on workflow and definition IDs
- Refactor services and modules to accommodate for new functionality

(your code organization is so tangled, it could be mistaken for a Christmas light display)
